Moving lighthouse code to a new `extensions` directory. Also added a `misc_utils.py` file.
